rename script and add comment
